Rework calls
We need the recent -S (status) command to probe the real state of
callaudiod.

- Lot of cleanup and refactorisation

Most of the code has been moved to a better place. The call audio
setups, the pickup action and the incall menu now are three different
things we can call individually.

- Handle failures individually

Every important task as the mmcli command or volume settings should warn
a urgent message to the user if they failed.

If we failed to setup audio, or to pickup the call then we dont even try
to open the incall menu.

- Made the incall_menu closeable and re-opennable

You can close the incoming call menu and the incall menu and re-open
them from any menu.

To make this possible I moved some of the responsibilities to the
modem monitor that trigger action after modem manager signals. We then
check at this point if it was the last call and do some cleanup if so.

- Initial work for concurrent calls

Added some code to the hooks and the modem checkfinished and
checkincoming calls to handle those cased. If you are in a call and someone
else try to call you, we refresh the incall menu and new entries will allow
you to switch calls.

- The incall menu isnt sticked to one call

This menu itself doesnt need a CALLID argument. It allow us to
manage every active call. We should be able to hold and hangup and
switch calls as we want.

+++ configs/default_hooks/sxmo_hook_missed_call.sh
@@ -9,6 +9,8 @@
 # $1 = Contact Name or Number (if not in contacts)
 
 # kill existing ring playback
-MPVID="$(cat "$XDG_RUNTIME_DIR/sxmo.ring.pid")"
-kill $MPVID
-rm "$XDG_RUNTIME_DIR/sxmo.ring.pid"
+if [ -e "$XDG_RUNTIME_DIR/sxmo.ring.pid" ]; then
+	MPVID="$(cat "$XDG_RUNTIME_DIR/sxmo.ring.pid")"
+	kill "$MPVID"
+	rm "$XDG_RUNTIME_DIR/sxmo.ring.pid"
+fi
