+++ configs/default_hooks/sxmo_hook_pickup.sh
@@ -5,7 +5,9 @@
 # This script is executed (asynchronously) when you pick up an incoming call
 
 # kill existing ring playback
-MPVID="$(cat "$XDG_RUNTIME_DIR/sxmo.ring.pid")"
-kill $MPVID
-rm "$XDG_RUNTIME_DIR/sxmo.ring.pid"
+if [ -e "$XDG_RUNTIME_DIR/sxmo.ring.pid" ]; then
+	MPVID="$(cat "$XDG_RUNTIME_DIR/sxmo.ring.pid")"
+	kill "$MPVID"
+	rm "$XDG_RUNTIME_DIR/sxmo.ring.pid"
+fi
